requirements.txt
appdirs==1.4.4
asgiref==3.3.0
attrs==20.3.0
certifi==2020.6.20
chardet==3.0.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.5.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
packaging==20.4
pipenv==2020.11.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytz==2020.4
requests==2.24.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.11
virtualenv==20.1.0
virtualenv-clone==0.5.4
whitenoise==5.2.0
